refactor(ext2): split the driver into one unit per concern (#318)
kernel/src/fs/ext2.c held 4412 lines and 98 functions, from
uuid_to_string to ext2_mount, with four functions over 100 lines. It was
the largest file in the kernel and the only way to read any part of the
driver.

The cost was not theoretical. #242 corrected the block-boundary
arithmetic in ext2_read_inode_data and left the identical expression in
ext2_write_inode_data alone; that surviving half is #311. The two
functions are near-identical and sit 80 lines apart inside 4412, which
is the condition under which a reader fixes one and believes both are
done. The file also already carried its own section banners, so the
decomposition existed already; it was never expressed on disk.

The driver is now kernel/src/fs/ext2/, split along the axis Linux uses
in fs/ext2 and following the shape kernel/src/io/video already has here,
a directory of units plus a private header:

    ext2.c            mount, the VFS operation tables, init and finalize
    ext2_io.c         block device to memory: superblock, block, BGDT, inode
    ext2_alloc.c      finding, allocating and releasing inodes and blocks
    ext2_blockmap.c   file block to device block, and inode data I/O
    ext2_dir.c        directory entries: iteration, creation, removal
    ext2_namei.c      resolving a path, and creating or removing a name
    ext2_file.c       operations on an open file
    ext2_attr.c       stat, statfs and setattr
    ext2_debug.c      dumps of the on-disk structures
    ext2_internal.h   the definitions the units share

No behaviour changes, and it is verifiable rather than asserted: all 98
function bodies are identical to the ones they replace, checked by
extracting each function from the old file and from the new tree and
comparing them. The only edits are the ones linkage forces:

  - 58 functions lose `static` and gain a declaration in the header,
    because they are now called across units; the other 29 stay static,
    and -Wall -Werror proves none of them became unreachable.
  - eleven small pure helpers (bitmap bit twiddling, record lengths,
    group index arithmetic) move into the header as `static inline`.
  - the two VFS operation tables lose `static`: ext2_init_vfs_file needs
    them, and it lives in ext2_namei.c.
  - the trace counter becomes `ext2_resource_id`, defined in ext2.c and
    declared in the header, because its two users are now in different
    units. EXT2_FULL_DEBUG moves to the header for the same reason: six
    units test it.
  - one signature now fits on a single line without `static` in front.

The public surface is unchanged: the two functions of fs/ext2.h.
CMakeLists.txt needs no change either, since the kernel sources are
collected with GLOB_RECURSE ... CONFIGURE_DEPENDS.

Verified with Debug and with -DENABLE_EXT2_TRACE=ON, both clean under
-Wall -Werror, and the full suite: 61 tests, 0 failures, 0 panics.

The maintainer documents that cited ext2.c by line number now name the
unit instead. The serial output quoted in ext2.md is left as it is: it
records a past run, and rewriting it would falsify the evidence.
